Crate csound

Csound

This crate contains safe Csound bindings for the csound's library. The supported csound's version is >= 6.12

Whats is Csound?

Csound is a sound and music computing system. If you want to known more visit:

Csound webside

Documentation

Community

Audio examples

Floss

Hello World

A simple Hello world example which will generate and reproduce a simple sine wave signal. A callback for catch messages from csound and print it into the string is configured. The call to the csound's perform() method will block the application until the end of the score have been reached. There are another alternatives for non blocking calls to perform csound's scores or csd files. see the examples in the project's source directory or go to csound's examples repository for more advanced examples and use cases.

extern crate csound;
use csound::*;

static score: &str = "<CsoundSynthesizer>
<CsOptions>
-odac
</CsOptions>
<CsInstruments>

sr = 44100
ksmps = 32
nchnls = 2
0dbfs  = 1

instr 1

kamp = .6
kcps = 440
ifn  = p4

asig oscil kamp, kcps, ifn
     outs asig,asig

endin
</CsInstruments>
<CsScore>
f1 0 16384 10 1
i 1 0 2 1
e
</CsScore>
</CsoundSynthesizer>";

fn main() {
    let mut cs = Csound::new();

   let func = |_, message:&str| {
       print!("{}", message);
   };
   cs.message_string_callback(func);
   cs.compile_csd_text(csd).unwrap();
   cs.start().unwrap();

   cs.perform();
}

Structs

CS_AudioDevice
Struct with specific audio device information.

CS_MidiDevice
Struct with specific MIDI device information.

ChannelHints
Holds the channel HINTS information.

ChannelInfo
Holds all relevant information about a csound bus channel.

CircularBuffer
Csound's Circular Buffer refresentation.

ControlChannelPtr
Rust representation for an raw csound channel pointer

ControlChannelType
Defines the types of csound bus cahnnels

Csound
Opaque struct representing an csound object

FileInfo
Struct containing the relevant info of files opened by csound.

OpcodeListEntry
Struct with information about a csound opcode.

RT_AudioParams
Real time audio params for a specific audio Device.

Table
Csound table representation.

pvs_DataExt
Holds pvs data info of a pvs channel.

Enums

ChannelData
Enum variant which represent the types of channels.

FileTypes
Describes the differents file types supported by csound

Language
ENums contains the supported csound languages

MessageType
Define the type of csound messages

Status
Csound error codes

Type Definitions

RTCLOCK