Struct csound::BufferPtr

pub struct BufferPtr<'a, T> { /* fields omitted */ }

Csound buffer pointer representation. This struct is build up to manipulate directly csound's buffers.

Methods

impl<'a, T> BufferPtr<'a, T>

pub fn get_size(&self) -> usize

Returns

The buffer length

pub fn copy_to_slice(&self, slice: &mut [f64]) -> usize

This method is used to copy data from the csound's buffer into another slice.

Arguments

slice A mutable slice where the data will be copy

Returns

The number of elements copied into the slice.

pub fn as_slice(&self) -> &[f64]

Returns

A slice to the buffer internal data

impl<'a> BufferPtr<'a, Writable>

pub fn as_mut_slice(&mut self) -> &mut [f64]

Returns

This buffer pointer as a mutable slice.

pub fn copy_from_slice(&self, slice: &[f64]) -> usize

method used to copy data into this buffer

Arguments

slice A slice with samples to copy

Returns

The number of elements copied into the csound's buffer.